+++ src/compojure/api/exception.clj
@@ -56,6 +56,11 @@
         (update :coercion cc/get-name)
         (->> (cc/encode-error (:coercion data))))))
 
+(defn http-response-handler
+  "reads response from ex-data :response"
+  [_ {:keys [response]} _]
+  response)
+
 (defn schema-error-handler
   "Creates error response based on Schema error."
   [e data req]